JavaScript Strings Homework Tasks
1. Write a JavaScript function reverses string and returns it
Example: "sample"  "elpmas".
2. Write a JavaScript function to check if in a given expression the
brackets are put correctly.
3. Write a JavaScript function that finds how many times a substring is
contained in a given text (perform case insensitive search).
4. You are given a text. Write a function that changes the text in all
regions:
• <upcase>text</upcase> to uppercase.
• <lowcase>text</lowcase> to lowercase
• <mixcase>text</mixcase> to mix casing(random)
5. Write a function that replaces non breaking white-spaces in a text
with &nbsp;
6. Write a function that extracts the content of a html page given as
text. The function should return anything that is in a tag, without the
tags
7. Write a script that parses an URL address given in the format:
[protocol]://[server]/[resource] and extracts from it the [protocol],
[server] and [resource] elements. Return the elements in a JSON object.
8. Write a JavaScript function that replaces in a HTML document given as
string all the tags <a href="…">…</a> with corresponding tags
[URL=…]…/URL].
9. Write a function for extracting all email addresses from given text.
All substrings that match the format <identifier>@<host>…<domain> should
be recognized as emails. Return the emails as array of strings.
10. Write a program that extracts from a given text all palindromes,
e.g. "ABBA", "lamal", "exe".
11. Write a function that formats a string using placeholders:
var str = stringFormat("Hello {0}!","Peter");
The function should work with up to 30 placeholders and all types
12. Write a function that creates a HTML UL using a template for every
HTML LI. The source of the list should an array of elements. Replace all
placeholders marked with –{…}–   with the value of the corresponding
property of the object.